requirements.txt
asgiref==3.5.0
boto3==1.21.34
botocore==1.24.34
Django==4.0.2
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.0
Pillow==9.0.1
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9

